refactor(health): split oversized modules
Organizational polish pass. Pure relocation — no behavior change,
public API unchanged.

health/mod.rs (732 \u2192 69 lines)
  \u2192 compute.rs (522)  \u2014 compute_health, compute_single_check, store_report,
                           compute_link_check, apply_dismissals, BrokenLinkItem
  \u2192 run_fix.rs (173)  \u2014 run_fix orchestrator
  mod.rs now holds module decls, re-exports, and the tiny pair_key helper.

health/custom.rs (1347 lines) \u2192 health/custom/ directory
  \u2192 mod.rs (183)      \u2014 orchestrator (run_all, preview_rule), evaluate
                           dispatch, finalize, module re-exports
  \u2192 types.rs (160)    \u2014 CustomRule, DomainMatchMode, CustomCheck,
                           PreviewResult, RawOutcome, FlaggedAtom
  \u2192 helpers.rs (158)  \u2014 shared preview/word_count/count_citations/host_of/
                           host_matches/for_each_atom/push_flag/load_candidates
  \u2192 rules.rs (472)    \u2014 all 11 eval_* fns (one per CustomRule variant)
  \u2192 tests.rs (412)    \u2014 existing #[cfg(test)] block moved verbatim

HealthPanel.tsx (849 \u2192 534 lines)
  \u2192 healthPanel.helpers.ts (328) \u2014 pure helpers + constants + filter types
                                      (HealthReport, CHECK_LABELS,
                                      CHECK_DESCRIPTIONS, FIX_ACTION_LABELS,
                                      STATUS_COLORS, CHECK_ORDER,
                                      DEFAULT_FILTER, SeverityFilter/
                                      FixableFilter/SortOrder, FilterState,
                                      pendingActions, manualOnlyCategories,
                                      extractExamples, extractCount,
                                      reviewItems, getSeverityBadge,
                                      getVisibleChecks)
  \u2192 ScoreBar.tsx (15)            \u2014 small presentational score meter
  HealthPanel.tsx now holds state, effects, keyboard shortcuts, and the
  JSX for the main panel \u2014 no more pure-helper soup at the top.

Deferred (not in this pass):
  - llm_fixes.rs (1255 lines) \u2014 split by action kind (merge/contradictions/
    overlap/tagging/tag-tree/boilerplate/links). Cohesive theme, already
    banner-organized, fine-grained public API; risky to split in a polish
    commit. Candidate for a v2 pass.
  - Crate-root audit \u2014 the only health-PR addition at atomic-core/src/
    root is boilerplate.rs, which is a cross-cutting content-fingerprinting
    utility shared by embedding, chunks, storage, and health. Crate root
    is the correct location.

Verification:
  cargo check  -p atomic-core -p atomic-server                       clean
  cargo check  -p atomic-core -p atomic-server --features postgres   clean
  cargo test   -p atomic-core -p atomic-server                       483 passed
  npx tsc --noEmit                                                   clean
  npx vitest run                                                     115 / 115
